agent-chaos/pyproject.toml
[project]
name = "agent-chaos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "discord-py>=2.6.4",
    "openai>=2.15.0",
    "paramiko>=4.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
